+++ limosat/__init__.py
@@ -1,5 +1,25 @@
-# limosat/__init__.py
+# MIT License
+#
+# Copyright (c) 2025 Sean Minhui Tashi Chua, and Anton Korosov
+#
+# Licensed under the MIT License. See the LICENSE file in the project root for full details.
 
-from .limosat import Keypoints, ImageProcessor, extract_templates, compute_descriptors, Matcher, KeypointDetector, Image
+"""
+LiMOSAT: A Python package for sea ice drift analysis.
+"""
 
-from .utils import logger, log_execution_time, extract_date
+from .image_processor import ImageProcessor
+from .image import Image
+from .keypoints import Keypoints
+from .keypoint_detector import KeypointDetector
+from .matcher import Matcher
+from .templates import Templates
+
+__all__ = [
+    "ImageProcessor",
+    "Image",
+    "Keypoints",
+    "KeypointDetector",
+    "Matcher",
+    "Templates",
+]
